Revert "Refactor transform_errors into middleware module (#26)"
This reverts commit 8f5fde70d219449b623d6c17bc86f0dc03eb80b5.
