+++ lib/plate_slate_web/resolvers/menu.ex
@@ -15,9 +15,25 @@ defmodule PlateSlateWeb.Resolvers.Menu do
   end
 
   def create_item(_, %{input: params}, _) do
-      with {:ok, item} <- Menu.create_item(params) do
-        {:ok, %{menu_item: item}}
+    case Menu.create_item(params) do
+      {:error, changeset} ->
+        {:ok, %{errors: transform_errors(changeset)}}
+
+      {:ok, menuitem} ->
+        {:ok, %{menu_item: menuitem}}
     end
   end
 
+  defp transform_errors(changeset) do
+    changeset
+    |> Ecto.Changeset.traverse_errors(&format_error/1)
+    |> Enum.map(fn {key, value} -> %{key: key, message: value} end)
+  end
+
+  @spec format_error(Ecto.Changeset.error()) :: String.t()
+  defp format_error({msg, opts}) do
+    Enum.reduce(opts, msg, fn {key, value}, acc ->
+      String.replace(acc, "%{#{key}}", to_string(value))
+    end)
+  end
 end
