--- lib/plate_slate_web/schema/middleware/changeset_errors.ex
@@ -1,26 +1,0 @@
-
-defmodule PlateSlateWeb.Schema.Middleware.ChangesetErrors do
-  @behaviour Absinthe.Middleware
-
-  def call(res, _) do
-    with %{errors: [%Ecto.Changeset{} = changeset]} <- res do
-      %{res |
-        value: %{errors: transform_errors(changeset)},
-        errors: []
-      }
-    end
-  end
-
-  defp transform_errors(changeset) do
-    changeset
-    |> Ecto.Changeset.traverse_errors(&format_error/1)
-    |> Enum.map(fn {key, value} -> %{key: key, message: value} end)
-  end
-
-  @spec format_error(Ecto.Changeset.error()) :: String.t()
-  defp format_error({msg, opts}) do
-    Enum.reduce(opts, msg, fn {key, value}, acc ->
-      String.replace(acc, "%{#{key}}", to_string(value))
-    end)
-  end
-end
